Move bpmn logic to separate file
